+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [8.1.1] - 2021-06-25
+### Fixed:
+- Handles `Uncaught ReferenceError: process is not defined` during getting if testing or not.
+
 ## [8.1.0] - 2021-06-25
 ### Added:
 - `SESSION_CREATED` event, which can be consumed by `onHandleEvent`